Changed the Makefile to use UNIX commands and fixed a few warnings and TODOs
Fixed warnings in pageFault.c and process.c
Implemented the suspension code
Can't be bothered with some signals flags :)

+++ src/memmgr/pageFault.c
@@ -83,7 +83,7 @@ void MemPageFaultHandler(IntrContext * intContext)
 	else
 	{
 		//Kernel mode fault
-		if(faultAddress < 4096)
+		if((unsigned int) faultAddress < 4096)
 		{
 			Panic("MemPageFaultHandler: Unable to handle kernel NULL pointer dereference");
 		}
